pyproject.toml
[project]
name = "aw1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.1.0",
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
]
